+++ pico-cnn/utils.c
@@ -1,7 +1,7 @@
 #include "utils.h"
 
 void extend_2d_input_with_padding(const fp_t* input_channel, const uint16_t height, const uint16_t width,
-                                  fp_t** extended_input, const int* padding, fp_t initializer) {
+                                  fp_t** extended_input, const uint16_t* padding, fp_t initializer) {
 
     uint16_t height_padded = height + padding[0] + padding[2];
     uint16_t width_padded = width + padding[1] + padding[3];
@@ -9,8 +9,8 @@ void extend_2d_input_with_padding(const fp_t* input_channel, const uint16_t heig
     *extended_input = (fp_t*) calloc(height_padded * width_padded, sizeof(fp_t));
 
     if(initializer != 0.0) {
-        for(int r = 0; r < height_padded; r++){
-            for(int c = 0; c < width_padded; c++){
+        for(uint16_t r = 0; r < height_padded; r++){
+            for(uint16_t c = 0; c < width_padded; c++){
                 (*extended_input)[r*width_padded+c] = initializer;
             }
         }
@@ -23,14 +23,14 @@ void extend_2d_input_with_padding(const fp_t* input_channel, const uint16_t heig
 }
 
 void extend_1d_input_with_padding(const fp_t* input_channel, const uint16_t width,
-                                  fp_t** extended_input, const int* padding, fp_t initializer) {
+                                  fp_t** extended_input, const uint16_t* padding, fp_t initializer) {
 
     uint16_t width_padded = width + padding[0] + padding[1];
 
     *extended_input = (fp_t*) calloc(width_padded, sizeof(fp_t));
 
     if(initializer != 0.0) {
-        for(int i = 0; i < width_padded; i++){
+        for(uint16_t i = 0; i < width_padded; i++){
             (*extended_input)[i] = initializer;
         }
     }
